Some small correction was made to query data from cities and states. VERY IMPORTANT:
Warning

This extension was deprecated in PHP 5.5.0, and it was removed in PHP 7.0.0. Instead, the MySQLi or PDO_MySQL extension should be used. See also MySQL: choosing an API guide and related FAQ for more information. Alternatives to this function include:

    mysqli_query()
    PDO::query()

It need to be changed as soon as possible. The preference here is to use PDO. 
All queries are located in the `includes/Sql/sql.class.php`